Refactored code with Stef
