+++ .env-sample
@@ -7,3 +7,5 @@ DB_USER=<user_name_here>
 DB_SECRET=<db_secret_here>
 DB_DATABASE=<db_name_here>
 DB_PORT=<db_port_here>
+
+JWT_SECRET=<Your-secret-here>
