+++ REST/crud_teachers.rest
@@ -1,9 +1,9 @@
 
-GET http://localhost:8080/all
+GET http://localhost:8080/sql/teachers/all
 Content-Type: application/json
 
 ###
-POST http://localhost:8080/createNew
+POST http://localhost:8080/sql/teachers/createNew
 Content-Type: application/json
 
 {
@@ -15,11 +15,11 @@ Content-Type: application/json
 }
 
 ###
-GET http://localhost:8080/findOne/6
+GET http://localhost:8080/sql/teachers/findOne/2
 Content-Type: application/json
 
 ###
-POST http://localhost:8080/updateTeacher/6
+POST http://localhost:8080/sql/teachers/updateOne/2
 Content-Type: application/json
 
 {
@@ -31,5 +31,6 @@ Content-Type: application/json
 }
 
 ###
-POST http://localhost:8080/delete/1
+POST http://localhost:8080/sql/teachers/delete/3
 Content-Type: application/json
+
