+++ REST/register_request.rest
@@ -1,5 +1,5 @@
 ###POST 
-POST http://localhost:8080/register/ 
+POST http://localhost:8080/sql/register/ 
 Content-Type: application/json
 
 {
