--- REST/update_request.rest
@@ -1,15 +1,0 @@
-POST http://localhost:8080/update/5
-Content-Type: application/json
-
-{
-    "first_name": "Not",
-    "last_name": "Anders",
-    "email": "teste@test.com",
-    "password": "qweqweqwe",
-    "department_id": 55
-}
-
-###
-
-GET http://localhost:8080/update/5
-Content-Type: application/json
