+++ connectors/db.mysql.js
@@ -1,8 +1,6 @@
 const { Sequelize } = require('sequelize');
 const { initModels } = require('../models/init-models.js')
 
-console.log("here")
-
 // define mysql database connection
 const sequelize = new Sequelize({
     host: process.env.DB_HOST,
@@ -12,19 +10,8 @@ const sequelize = new Sequelize({
     database: process.env.DB_NAME,
     dialect: 'mysql'
 });
+ 
 
-
-//Synchronizes the models each time the application is started
-//sequelize.sync(); 
-
-(async () => {
-    try {
-        await sequelize.authenticate();
-        console.log('Connection to DB established successfully');
-    } catch (error) {
-        console.log(`Unable to connect to DB: ${error}`);
-    }
-});
 // define sequelize models
 initModels(sequelize);
